test(cypress): reorganize parallel spec files
