+++ integration-tests/cypress/cypress-final-status-impacted-tests.spec.js
@@ -1,13 +1,17 @@
 'use strict'
 
 const assert = require('node:assert/strict')
-const { exec } = require('node:child_process')
+const { exec, execSync } = require('node:child_process')
 const { once } = require('node:events')
+const fs = require('node:fs')
+const path = require('node:path')
 
+const semver = require('semver')
 const {
   sandboxCwd,
   useSandbox,
   getCiVisEvpProxyConfig,
+  assertObjectContains,
   stopCiVisTestEnv,
   warmCypressBinary,
 } = require('../helpers')
@@ -22,6 +26,20 @@ const {
   TEST_MANAGEMENT_IS_DISABLED,
   TEST_NAME,
   TEST_FINAL_STATUS,
+  TEST_SOURCE_FILE,
+  TEST_EARLY_FLAKE_ENABLED,
+  TEST_SESSION_NAME,
+  TEST_RETRY_REASON,
+  DD_CAPABILITIES_TEST_IMPACT_ANALYSIS,
+  DD_CAPABILITIES_EARLY_FLAKE_DETECTION,
+  DD_CAPABILITIES_AUTO_TEST_RETRIES,
+  DD_CAPABILITIES_IMPACTED_TESTS,
+  DD_CAPABILITIES_TEST_MANAGEMENT_QUARANTINE,
+  DD_CAPABILITIES_TEST_MANAGEMENT_DISABLE,
+  DD_CAPABILITIES_TEST_MANAGEMENT_ATTEMPT_TO_FIX,
+  DD_CAPABILITIES_FAILED_TEST_REPLAY,
+  TEST_RETRY_REASON_TYPES,
+  TEST_IS_MODIFIED,
 } = require('../../packages/dd-trace/src/plugins/util/test')
 const { DD_MAJOR, NODE_MAJOR } = require('../../version')
 
@@ -29,6 +47,8 @@ const requestedVersion = process.env.CYPRESS_VERSION
 const oldestVersion = DD_MAJOR >= 6 ? '12.0.0' : '6.7.0'
 const version = requestedVersion === 'oldest' ? oldestVersion : requestedVersion
 const hookFile = 'dd-trace/loader-hook.mjs'
+const NUM_RETRIES_EFD = 3
+const over12It = (version === 'latest' || semver.gte(version, '12.0.0')) ? it : it.skip
 
 function shouldTestsRun (type) {
   if (DD_MAJOR === 5) {
@@ -558,6 +578,417 @@ moduleTypes.forEach(({
           once(childProcess, 'exit'),
           eventsPromise,
         ])
+      })
+    })
+
+    context('libraries capabilities', () => {
+      it('adds capabilities to tests', async () => {
+        const envVars = getCiVisEvpProxyConfig(receiver.port)
+
+        const specToRun = 'cypress/e2e/spec.cy.js'
+
+        childProcess = exec(
+          version === 'latest' ? testCommand : `${testCommand} --spec ${specToRun}`,
+          {
+            cwd,
+            env: {
+              ...envVars,
+              CYPRESS_BASE_URL: webAppBaseUrl,
+              DD_TEST_SESSION_NAME: 'my-test-session-name',
+              SPEC_PATTERN: specToRun,
+            },
+          }
+        )
+
+        const receiverPromise = receiver
+          .gatherPayloadsUntilChildExit(
+            childProcess,
+            ({ url }) => url.endsWith('/api/v2/citestcycle'),
+            payloads => {
+              const metadataDicts = payloads
+                .filter(({ payload }) => payload.metadata?.test)
+                .flatMap(({ payload }) => payload.metadata)
+
+              assert.ok(metadataDicts.length > 0, `Expected ${metadataDicts.length} > 0`)
+              metadataDicts.forEach(metadata => {
+                assert.strictEqual(metadata.test[DD_CAPABILITIES_TEST_IMPACT_ANALYSIS], '1')
+                assert.strictEqual(metadata.test[DD_CAPABILITIES_EARLY_FLAKE_DETECTION], '1')
+                assert.strictEqual(metadata.test[DD_CAPABILITIES_AUTO_TEST_RETRIES], '1')
+                assert.strictEqual(metadata.test[DD_CAPABILITIES_IMPACTED_TESTS], '1')
+                assert.strictEqual(metadata.test[DD_CAPABILITIES_TEST_MANAGEMENT_QUARANTINE], '1')
+                assert.strictEqual(metadata.test[DD_CAPABILITIES_TEST_MANAGEMENT_DISABLE], '1')
+                assert.strictEqual(metadata.test[DD_CAPABILITIES_TEST_MANAGEMENT_ATTEMPT_TO_FIX], '5')
+                assert.strictEqual(metadata.test[DD_CAPABILITIES_FAILED_TEST_REPLAY], '1')
+                // capabilities logic does not overwrite test session name
+                assert.strictEqual(metadata['*'][TEST_SESSION_NAME], 'my-test-session-name')
+              })
+            }, { hardTimeout: 25000 })
+
+        await Promise.all([
+          once(childProcess, 'exit'),
+          receiverPromise,
+        ])
+      })
+    })
+
+    context('impacted tests', () => {
+      beforeEach(() => {
+        receiver.setKnownTests({
+          cypress: {
+            'cypress/e2e/impacted-test.js': ['impacted test is impacted test'],
+          },
+        })
+      })
+
+      // Add git setup before running impacted tests
+      before(function () {
+        execSync('git checkout -b feature-branch', { cwd, stdio: 'ignore' })
+        fs.writeFileSync(
+          path.join(cwd, 'cypress/e2e/impacted-test.js'),
+          `/* eslint-disable */
+          describe('impacted test', () => {
+            it('is impacted test', () => {
+              cy.visit('/')
+                .get('.hello-world')
+                .should('have.text', 'Hello Worldd')
+            })
+          })`
+        )
+        execSync('git add cypress/e2e/impacted-test.js', { cwd, stdio: 'ignore' })
+        execSync('git commit -m "modify impacted-test.js"', { cwd, stdio: 'ignore' })
+
+        // Modify impacted-test-order.js to make it "impacted"
+        const currentContent = fs.readFileSync(path.join(cwd, 'cypress/e2e/impacted-test-order.js'), 'utf-8')
+        fs.writeFileSync(
+          path.join(cwd, 'cypress/e2e/impacted-test-order.js'),
+          currentContent + '\n// modified'
+        )
+        execSync('git add cypress/e2e/impacted-test-order.js', { cwd, stdio: 'ignore' })
+        execSync('git commit -m "modify impacted-test-order.js"', { cwd, stdio: 'ignore' })
+      })
+
+      after(function () {
+        execSync('git checkout -', { cwd, stdio: 'ignore' })
+        execSync('git branch -D feature-branch', { cwd, stdio: 'ignore' })
+      })
+
+      const getTestAssertions = ({ isModified, isEfd, isNew }, childProcess) =>
+        receiver
+          .gatherPayloadsUntilChildExit(
+            childProcess,
+            ({ url }) => url.endsWith('/api/v2/citestcycle'),
+            (payloads) => {
+              const events = payloads.flatMap(({ payload }) => payload.events)
+              const tests = events.filter(event => event.type === 'test').map(event => event.content)
+              const testSession = events.find(event => event.type === 'test_session_end').content
+
+              if (isEfd) {
+                assert.strictEqual(testSession.meta[TEST_EARLY_FLAKE_ENABLED], 'true')
+              } else {
+                assert.ok(!(TEST_EARLY_FLAKE_ENABLED in testSession.meta))
+              }
+
+              const resourceNames = tests.map(span => span.resource)
+
+              assertObjectContains(resourceNames,
+                [
+                  'cypress/e2e/impacted-test.js.impacted test is impacted test',
+                ]
+              )
+
+              const impactedTests = tests.filter(test =>
+                test.meta[TEST_SOURCE_FILE] === 'cypress/e2e/impacted-test.js' &&
+              test.meta[TEST_NAME] === 'impacted test is impacted test')
+
+              if (isEfd) {
+                assert.strictEqual(impactedTests.length, NUM_RETRIES_EFD + 1) // Retries + original test
+              } else {
+                assert.strictEqual(impactedTests.length, 1)
+              }
+
+              for (const impactedTest of impactedTests) {
+                if (isModified) {
+                  assert.strictEqual(impactedTest.meta[TEST_IS_MODIFIED], 'true')
+                } else {
+                  assert.ok(!(TEST_IS_MODIFIED in impactedTest.meta))
+                }
+                if (isNew) {
+                  assert.strictEqual(impactedTest.meta[TEST_IS_NEW], 'true')
+                } else {
+                  assert.ok(!(TEST_IS_NEW in impactedTest.meta))
+                }
+              }
+
+              if (isEfd) {
+                const retriedTests = tests.filter(
+                  test => test.meta[TEST_IS_RETRY] === 'true' &&
+                test.meta[TEST_NAME] === 'impacted test is impacted test'
+                )
+                assert.strictEqual(retriedTests.length, NUM_RETRIES_EFD)
+                let retriedTestNew = 0
+                let retriedTestsWithReason = 0
+                retriedTests.forEach(test => {
+                  if (test.meta[TEST_IS_NEW] === 'true') {
+                    retriedTestNew++
+                  }
+                  if (test.meta[TEST_RETRY_REASON] === TEST_RETRY_REASON_TYPES.efd) {
+                    retriedTestsWithReason++
+                  }
+                })
+                assert.strictEqual(retriedTestNew, isNew ? NUM_RETRIES_EFD : 0)
+                assert.strictEqual(retriedTestsWithReason, NUM_RETRIES_EFD)
+              }
+            }, 25000)
+
+      const runImpactedTest = async (
+        { isModified, isEfd = false, isNew = false },
+        extraEnvVars = {}
+      ) => {
+        const envVars = getCiVisEvpProxyConfig(receiver.port)
+
+        const specToRun = 'cypress/e2e/impacted-test.js'
+
+        childProcess = exec(
+          version === 'latest' ? testCommand : `${testCommand} --spec ${specToRun}`,
+          {
+            cwd,
+            env: {
+              ...envVars,
+              CYPRESS_BASE_URL: webAppBaseUrl,
+              SPEC_PATTERN: specToRun,
+              GITHUB_BASE_REF: '',
+              ...extraEnvVars,
+            },
+          }
+        )
+
+        const testAssertionsPromise = getTestAssertions({ isModified, isEfd, isNew }, childProcess)
+
+        await Promise.all([
+          once(childProcess, 'exit'),
+          testAssertionsPromise,
+        ])
+      }
+
+      context('test is not new', () => {
+        it('should be detected as impacted', async () => {
+          receiver.setSettings({ impacted_tests_enabled: true })
+
+          await runImpactedTest({ isModified: true })
+        })
+
+        it('should not be detected as impacted if disabled', async () => {
+          receiver.setSettings({ impacted_tests_enabled: false })
+
+          await runImpactedTest({ isModified: false })
+        })
+
+        it('should not be detected as impacted if DD_CIVISIBILITY_IMPACTED_TESTS_DETECTION_ENABLED is false',
+          async () => {
+            receiver.setSettings({ impacted_tests_enabled: true })
+
+            await runImpactedTest(
+              { isModified: false },
+              { DD_CIVISIBILITY_IMPACTED_TESTS_DETECTION_ENABLED: '0' }
+            )
+          })
+      })
+
+      context('test is new', () => {
+        it('should be retried and marked both as new and modified', async () => {
+          receiver.setKnownTests({
+            cypress: {},
+          })
+          receiver.setSettings({
+            impacted_tests_enabled: true,
+            early_flake_detection: {
+              enabled: true,
+              slow_test_retries: {
+                '5s': NUM_RETRIES_EFD,
+              },
+            },
+            known_tests_enabled: true,
+          })
+          await runImpactedTest(
+            { isModified: true, isEfd: true, isNew: true }
+          )
+        })
+      })
+
+      over12It('does not retry impacted tests when testIsolation is false', async () => {
+        receiver.setSettings({
+          impacted_tests_enabled: true,
+          early_flake_detection: {
+            enabled: true,
+            slow_test_retries: {
+              '5s': NUM_RETRIES_EFD,
+            },
+          },
+          known_tests_enabled: true,
+        })
+
+        const envVars = getCiVisEvpProxyConfig(receiver.port)
+
+        const specToRun = 'cypress/e2e/impacted-test.js'
+
+        childProcess = exec(
+          version === 'latest' ? testCommand : `${testCommand} --spec ${specToRun}`,
+          {
+            cwd,
+            env: {
+              ...envVars,
+              CYPRESS_BASE_URL: webAppBaseUrl,
+              SPEC_PATTERN: specToRun,
+              GITHUB_BASE_REF: '',
+              CYPRESS_TEST_ISOLATION: 'false',
+            },
+          }
+        )
+
+        const receiverPromise = receiver
+          .gatherPayloadsUntilChildExit(
+            childProcess,
+            ({ url }) => url.endsWith('/api/v2/citestcycle'),
+            (payloads) => {
+              const events = payloads.flatMap(({ payload }) => payload.events)
+              const tests = events.filter(event => event.type === 'test').map(event => event.content)
+              const testSession = events.find(event => event.type === 'test_session_end').content
+
+              assertObjectContains(testSession.meta, {
+                [TEST_EARLY_FLAKE_ENABLED]: 'true',
+              })
+
+              const impactedTests = tests.filter(test =>
+                test.meta[TEST_SOURCE_FILE] === 'cypress/e2e/impacted-test.js' &&
+              test.meta[TEST_NAME] === 'impacted test is impacted test')
+
+              // Should only have 1 test, no retries when testIsolation is false
+              assert.equal(impactedTests.length, 1)
+
+              for (const impactedTest of impactedTests) {
+                assertObjectContains(impactedTest.meta, {
+                  [TEST_IS_MODIFIED]: 'true',
+                })
+              }
+
+              // No retries should occur when testIsolation is false
+              const retriedTests = tests.filter(
+                test => test.meta[TEST_IS_RETRY] === 'true' &&
+              test.meta[TEST_NAME] === 'impacted test is impacted test'
+              )
+              assert.equal(retriedTests.length, 0)
+            }, { hardTimeout: 25000 })
+
+        await Promise.all([
+          once(childProcess, 'exit'),
+          receiverPromise,
+        ])
+      })
+
+      it('retries impacted tests in the correct order (right after original test)', async () => {
+        let testOutput = ''
+        receiver.setSettings({
+          impacted_tests_enabled: true,
+          early_flake_detection: {
+            enabled: true,
+            slow_test_retries: {
+              '5s': 2,
+            },
+          },
+          known_tests_enabled: true,
+        })
+
+        receiver.setKnownTests({
+          cypress: {
+            'cypress/e2e/impacted-test-order.js': [
+              'impacted test order first test',
+              'impacted test order second test',
+            ],
+          },
+        })
+
+        const envVars = getCiVisEvpProxyConfig(receiver.port)
+
+        const specToRun = 'cypress/e2e/impacted-test-order.js'
+
+        childProcess = exec(
+          version === 'latest' ? testCommand : `${testCommand} --spec ${specToRun}`,
+          {
+            cwd,
+            env: {
+              ...envVars,
+              CYPRESS_BASE_URL: webAppBaseUrl,
+              SPEC_PATTERN: specToRun,
+              GITHUB_BASE_REF: '',
+            },
+          }
+        )
+
+        const receiverPromise = receiver
+          .gatherPayloadsUntilChildExit(
+            childProcess,
+            ({ url }) => url.endsWith('/api/v2/citestcycle'),
+            (payloads) => {
+              const events = payloads.flatMap(({ payload }) => payload.events)
+              const tests = events.filter(event => event.type === 'test').map(event => event.content)
+
+              // All tests in the file are new and modified, so they should all be retried
+              // 2 tests * (1 original + 2 retries) = 6 tests total
+              assert.equal(tests.length, 6)
+
+              // Extract test execution order with full details
+              const testExecutionOrder = tests.map(test => ({
+                name: test.meta[TEST_NAME],
+                isRetry: test.meta[TEST_IS_RETRY] === 'true',
+                isModified: test.meta[TEST_IS_MODIFIED] === 'true',
+              }))
+
+              // All should be marked as modified
+              testExecutionOrder.forEach(test => {
+                assert.equal(test.isModified, true)
+              })
+
+              // Expected order:
+              // 1. "first test" (original)
+              // 2. "first test" (retry 1)
+              // 3. "first test" (retry 2)
+              // 4. "second test" (original)
+              // 5. "second test" (retry 1)
+              // 6. "second test" (retry 2)
+
+              assertObjectContains(testExecutionOrder, [
+                { name: 'impacted test order first test', isRetry: false },
+                { name: 'impacted test order first test', isRetry: true },
+                { name: 'impacted test order first test', isRetry: true },
+                { name: 'impacted test order second test', isRetry: false },
+                { name: 'impacted test order second test', isRetry: true },
+                { name: 'impacted test order second test', isRetry: true },
+              ])
+
+              const testSession = events.find(event => event.type === 'test_session_end').content
+              assertObjectContains(testSession.meta, {
+                [TEST_EARLY_FLAKE_ENABLED]: 'true',
+              })
+            }, { hardTimeout: 25000 })
+
+        childProcess.stdout?.on('data', (data) => {
+          testOutput += data.toString()
+          process.stdout.write(data)
+        })
+        childProcess.stderr?.on('data', (data) => {
+          testOutput += data.toString()
+          process.stderr.write(data)
+        })
+
+        await Promise.all([
+          once(childProcess, 'exit'),
+          once(childProcess.stdout, 'end'),
+          once(childProcess.stderr, 'end'),
+          receiverPromise,
+        ])
+
+        assert.match(testOutput, /Retrying "impacted test order first test" to detect flakes because it is modified/)
+        assert.match(testOutput, /Retrying "impacted test order second test" to detect flakes because it is modified/)
       })
     })
   })
